+++ app/src/main/java/com/mega/metaweather/domain/model/DisplayData.kt
@@ -1,9 +1,9 @@
 package com.mega.metaweather.domain.model
 
 data class DisplayData(
-    val lowestTemperature: Double,
-    val highestTemperature: Double,
-    val currentTemperature: Double,
+    val lowestTemperature: Int,
+    val highestTemperature: Int,
+    val currentTemperature: Int,
     val weather: String,
     val weather_abbr: String,
     val title: String,
